ok funcionando para presentacion viernes12

+++ Front/src/components/Navbar.jsx
@@ -1,14 +1,20 @@
 import React from 'react'
-import { Link } from 'react-router-dom'
+import { Link, useHistory } from 'react-router-dom'
 import Button from '@material-ui/core/Button';
 import themes from '../themes/themesConfig'
 import styles from '../styles/navbar.module.css'
 import {useDispatch, useSelector} from 'react-redux';
 import {queryWines} from '../store/wines';
 import {cleanUser} from "../store/user"
+import axios from "axios";
+import { productsArr } from "../store/productsArr"
+
 
 const Navbar = () => {
     const user = useSelector((state) => state.user)
+    const cart = useSelector((state) => state.cart)
+    const history = useHistory()
+
     const [query, setQuery] = React.useState('');
     const dispatch = useDispatch(); 
     const isLoggedIn = Object.keys(user).length 
@@ -21,15 +27,30 @@ const Navbar = () => {
     };    
 
     const logOutHandler = () => {
-        localStorage.clear()
-        dispatch(cleanUser())
+        if(isLoggedIn) {
+            const token = user.token
+            axios.put("http://localhost:5000/api/cart", cart, {
+                headers: { Authorization: `Bearer ${token}` }
+            })
+            .then(({data}) => data)
+            
+            localStorage.clear()
+            dispatch(cleanUser())
+            dispatch(productsArr([]))
+
+            history.push("/")
+        }else{
+            history.push("/login")
+        }  
     }
 
     return (
         <div className='barra'>
             <div className={styles.barraD}>
                 <div className={styles.orden}>
+                    <Link to="/">
                     <img src="https://www.ubp.edu.ar/wp-content/uploads/2014/09/logo-de-geovinar-pin-01.png" alt="" />
+                    </Link>
                     <div style={{display: 'flex'}}>
                         <input type="text" placeholder="Busca productos marcas y mas..." value={query} onChange={e => setQuery(e.target.value)}/>
                         <Button variant="outlined" color='primary' style={{height: '2.7em', margin: '0.9em 0 0 0.4em'}} onClick={() => handleQuery(query)}>Search</Button>
@@ -37,9 +58,9 @@ const Navbar = () => {
                 </div>
                 <div className={styles.barraI}>
                     <div className={styles.botonLogIn}>
-                        <Link to={isLoggedIn ? "/" : "/login"} onClick={logOutHandler}>
+                        <div onClick={logOutHandler}>
                             <Button variant="outlined" color='primary' >  { isLoggedIn ? "Log Out" : "Sign In" } </Button>
-                        </Link>
+                        </div>
                     </div>
                     
                     <div className={styles.botonLogOut}>
@@ -51,12 +72,15 @@ const Navbar = () => {
                                 </Link>
                             ) 
                             : 
-                                null
+                            (
+                            <div className={styles.carrito}>
+                            <Link to='/cart'><span className="material-icons">shopping_cart</span></Link>
+                            </div>
+                            )
+                                    
                         }
                     </div>
-                    <div className={styles.carrito}>
-                        <Link to='/'><span className="material-icons">shopping_cart</span></Link>
-                    </div>
+                    
                 </div>
             </div>
         </div>
